+++ src/constants/actions/file.js
@@ -1,0 +1,3 @@
+export const SAVE = 'FILE_SAVE';
+export const LOAD = 'FILE_LOAD';
+export const EXPORT_SVG = 'FILE_EXPORT_SVG';