+++ src/containers/App.js
@@ -1,5 +1,10 @@
 import { connect } from 'react-redux';
 import { open } from '../actions/Dialog';
+import {
+    save as onSave,
+    load as onLoad,
+    exportSVG as onExportSVG,
+} from '../actions/file';
 import { App as Component } from '../components/App';
 import { mapActionCreatorsToProps } from './Container';
 
@@ -9,6 +14,9 @@ export const App = connect(
         edit: state.editor.edit,
     }),
     mapActionCreatorsToProps({
-        open,
+        onOpenConfig: () => open('config'),
+        onLoad,
+        onSave,
+        onExportSVG,
     })
 )(Component);