+++ src/middlewares/__tests__/test-file.js
@@ -1,0 +1,4 @@
+describe('middleware/file', () => {
+    jest.unmock('../file');
+    require('../file');
+});