+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Theme color.
 ### Fixed
 - Fix board data has not been saved.
+- Fix style of dialogs for mobile.
 
 ## [1.3.0] - 2016-03-10
 ### Changed