+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased]
+## [1.4.0] - 2016-03-20
 ### Added
 - Save/Load/Export as file. ([#13](https://github.com/ukatama/nekoboard/issues/13)
 ### Changed